+++ server/routes/index.ts
@@ -5,7 +5,7 @@ import type { Services } from '../services'
 import HomeController from './homeController'
 import bookingJourneyRoutes from './bookVisit'
 import paths from '../constants/paths'
-import BookingController from './bookingController'
+import BookingsController from './bookingsController'
 
 export default function routes(services: Services): Router {
   const router = Router()
@@ -13,7 +13,7 @@ export default function routes(services: Services): Router {
   const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
 
   const home = new HomeController(services.bookerService)
-  const booking = new BookingController(services.visitService)
+  const booking = new BookingsController(services.visitService)
 
   get(paths.HOME, home.view())
   get(paths.BOOKINGS.HOME, booking.view())
